+++ app/classes/shared/main_controller.py
@@ -246,7 +246,7 @@ class Controller:
         except:
             return {"percent": 0, "total_files": 0}
 
-    def create_api_server(self, data: dict):
+    def create_api_server(self, data: dict, user_id):
         server_fs_uuid = Helpers.create_uuid()
         new_server_path = os.path.join(self.helper.servers_dir, server_fs_uuid)
         backup_path = os.path.join(self.helper.backup_path, server_fs_uuid)
@@ -309,7 +309,9 @@ class Controller:
                 # TODO: Copy files from the zip file to the new server directory
                 server_file = create_data["jarfile"]
                 raise Exception("Not yet implemented")
-            _create_server_properties_if_needed(create_data["server_properties_port"])
+            _create_server_properties_if_needed(
+                create_data["server_properties_port"],
+            )
 
             min_mem = create_data["mem_min"]
             max_mem = create_data["mem_max"]
@@ -405,6 +407,7 @@ class Controller:
             server_log_file=log_location,
             server_stop=stop_command,
             server_port=monitoring_port,
+            created_by=user_id,
             server_host=monitoring_host,
             server_type=monitoring_type,
         )
@@ -431,6 +434,7 @@ class Controller:
         min_mem: int,
         max_mem: int,
         port: int,
+        user_id: int,
     ):
         server_id = Helpers.create_uuid()
         server_dir = os.path.join(self.helper.servers_dir, server_id)
@@ -491,6 +495,7 @@ class Controller:
             server_log_file,
             server_stop,
             port,
+            user_id,
             server_type="minecraft-java",
         )
 
@@ -526,6 +531,7 @@ class Controller:
         min_mem: int,
         max_mem: int,
         port: int,
+        user_id: int,
     ):
         server_id = Helpers.create_uuid()
         new_server_dir = os.path.join(self.helper.servers_dir, server_id)
@@ -567,6 +573,7 @@ class Controller:
             server_log_file,
             server_stop,
             port,
+            user_id,
             server_type="minecraft-java",
         )
         ServersController.set_import(new_id)
@@ -581,6 +588,7 @@ class Controller:
         min_mem: int,
         max_mem: int,
         port: int,
+        user_id: int,
     ):
         server_id = Helpers.create_uuid()
         new_server_dir = os.path.join(self.helper.servers_dir, server_id)
@@ -623,6 +631,7 @@ class Controller:
             server_log_file,
             server_stop,
             port,
+            user_id,
             server_type="minecraft-java",
         )
         ServersController.set_import(new_id)
@@ -636,7 +645,12 @@ class Controller:
     # **********************************************************************************
 
     def import_bedrock_server(
-        self, server_name: str, server_path: str, server_exe: str, port: int
+        self,
+        server_name: str,
+        server_path: str,
+        server_exe: str,
+        port: int,
+        user_id: int,
     ):
         server_id = Helpers.create_uuid()
         new_server_dir = os.path.join(self.helper.servers_dir, server_id)
@@ -671,6 +685,7 @@ class Controller:
             server_log_file,
             server_stop,
             port,
+            user_id,
             server_type="minecraft-bedrock",
         )
         ServersController.set_import(new_id)
@@ -680,7 +695,12 @@ class Controller:
         return new_id
 
     def import_bedrock_zip_server(
-        self, server_name: str, zip_path: str, server_exe: str, port: int
+        self,
+        server_name: str,
+        zip_path: str,
+        server_exe: str,
+        port: int,
+        user_id: int,
     ):
         server_id = Helpers.create_uuid()
         new_server_dir = os.path.join(self.helper.servers_dir, server_id)
@@ -715,6 +735,7 @@ class Controller:
             server_log_file,
             server_stop,
             port,
+            user_id,
             server_type="minecraft-bedrock",
         )
         self.import_helper.import_bedrock_zip_server(
@@ -760,6 +781,7 @@ class Controller:
         server_log_file: str,
         server_stop: str,
         server_port: int,
+        created_by: int,
         server_type: str,
         server_host: str = "127.0.0.1",
     ):
@@ -774,6 +796,7 @@ class Controller:
             server_log_file,
             server_stop,
             server_type,
+            created_by,
             server_port,
             server_host,
         )